+++ api/security_firewall.go
@@ -114,6 +114,9 @@ func (api *API) ReadFirewallSettings(ctx context.Context, instanceID int) ([]map
 	case 200:
 		tflog.Debug(ctx, fmt.Sprintf("data: %v", data))
 		return data, nil
+	case 404:
+		tflog.Warn(ctx, "Firewall not found")
+		return nil, nil
 	default:
 		return nil, fmt.Errorf("failed to read firewall settings, status=%d message=%s ",
 			response.StatusCode, failed)